init: hello world
This was pretty fun. I do not know why the devcontainer seems to have
a default high match policy. It actually seems to select from the
highest set of cores, but then actually hand them out to the ranks
from the lowest to high. I suspect that fluxion returns the
high set and then flux hands them out still lowest first.
So for asking for 2 cores in 0-7, we first select 6-7, but
then assign 0:6 and 1:7

Signed-off-by: vsoch <vsoch@users.noreply.github.com>